# openapi.yaml
openapi: 3.0.3
info:
  title: Modelia AI Studio API
  description: Full-stack AI image generation application with JWT authentication
  version: 1.0.0
  contact:
    name: Modelia AI Studio Support

servers:
  - url: http://localhost:5000
    description: Development server

tags:
  - name: Authentication
    description: User authentication endpoints
  - name: Generation
    description: AI image generation endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
      required:
        - id
        - username

    Generation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        prompt:
          type: string
        style:
          type: string
          enum: [realistic, artistic, abstract, cyberpunk]
        imageUrl:
          type: string
        uploadedImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - prompt
        - style
        - imageUrl
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
      required:
        - user
        - token

    Error:
      type: object
      properties:
        error:
          type: string
        retryable:
          type: boolean
      required:
        - error

paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                password:
                  type: string
                  minLength: 6
              required:
                - username
                - password
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login to existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/generate:
    post:
      tags:
        - Generation
      summary: Generate AI artwork
      description: |
        Generate AI artwork based on prompt and style. Simulates 20% error rate
        for "Model overloaded" scenarios. Supports image upload for reference.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  minLength: 1
                  maxLength: 1000
                style:
                  type: string
                  enum: [realistic, artistic, abstract, cyberpunk]
                image:
                  type: string
                  format: binary
                  description: Optional reference image (max 10MB)
              required:
                - prompt
                - style
      responses:
        '200':
          description: Generation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generation'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: AI models overloaded (20% chance, retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Our AI models are currently experiencing high demand. Please try again in a moment."
                retryable: true

  /api/generations:
    get:
      tags:
        - Generation
      summary: Get generation history
      description: Returns last 5 generations for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Generation history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Generation'
                maxItems: 5
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/generations/{id}:
    get:
      tags:
        - Generation
      summary: Get specific generation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Generation retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied (not owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'